+++ src/class_inspector/utils.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+import re
+
+import black
+import isort
+import libcst as cst
+
+
+def _is_dunder(item: str) -> bool:
+    return item.startswith("__") and item.endswith("__")
+
+
+def camel_to_snake(name: str) -> str:
+    s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", name)
+    s2 = re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1)
+    return s2.lower()
+
+
+def format_code_str(code_snippet: str) -> str:
+    return black.format_str(isort.code(code_snippet), mode=black.FileMode())
+
+
+def str_to_cst(code: str) -> cst.Module:
+    return cst.parse_module(code)
+
+
+def cst_to_str(node) -> str:
+    return cst.Module([]).code_for_node(node)
